oo-admin-repair changes
Model changes to handle node_removed flag on the gear

+++ controller/app/models/pending_ops/destroy_gear_op.rb
@@ -4,8 +4,10 @@ class DestroyGearOp < PendingAppOp
   field :group_instance_id, type: String
 
   def execute
+    result_io = ResultIO.new
     gear = get_gear()
-    result_io = gear.destroy_gear(true)
+    result_io = gear.destroy_gear(true) unless gear.node_removed
+    result_io
   end
 
 end
